# Conflicts:
#	Core/Subscale/SubspaceCombiner/SequentialSubspaceCombiner.cpp

+++ Core/Subscale/CoreSetSeeker/CoreSet.h
@@ -12,7 +12,7 @@ public:
 	CoreSet();
 	CoreSet(Points);
 	void addPoint(Point* point);
-    int32_t getSize();
+    size_t getSize();
 };
 
 typedef std::vector<CoreSet> CoreSets;
